backend/pyproject.toml
[tool.poetry]
name = "ai-review"
version = "0.1.0"
description = "python libraries for ai-review"
authors = ["FilmBuachoom <filmbuachoom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.2.16"
psycopg2 = "2.9.9"
djangorestframework = "3.15.2"
django-environ = "0.11.2"
gunicorn = "22.0.0"
celery = "^5.4.0"
django-cors-headers = "^4.3.1"
numpy = "1.24.4"
pandas = "2.0.3"
nbib = "0.3.2"
requests = "2.32.3"
faiss-cpu = "1.7.4"
langchain-core = "0.3.35"
channels = "4.2.0"
channels-redis = "4.2.1"
daphne = "4.1.2"
twisted = {version = "^24.11.0", extras = ["http2", "tls"]}
langchain-google-genai = "2.0.9"
langchain-openai = "0.3.6"
google-api-python-client = "^2.169.0"
google-auth = "^2.40.1"

[tool.poetry.group.linux.dependencies]
tevatron = "0.1.0"
datasets = "1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
